+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fixed time filter updates not reflecting in the filter string.
 - Fixed index pattern changes not updating the browser URL.
 - Fixed container memory resource distribution
+- Fixed empty column status in cvs export
 ## New Features
 - Restricted time filter to a maximum range of one month.
 - Made the timestamp field mandatory and added a summary view when no fields are selected.